move code under the right file
